Move `filter` and `server` modules under `cli`.
