+++ tests/test-decks_test.py
@@ -2,7 +2,7 @@ import os
 import pytest
 
 from yanki.cli import find_errors
-from yanki.filter import DeckFilter, read_final_decks
+from yanki.cli.filter import DeckFilter, read_final_decks
 from yanki.video import VideoOptions
